ast: Extract ExprHolderNode class
This also replaces the AbstractExprStmtNode class.
